Remove ccresponse.cpp/.hpp files and move implementations of response related functions to sr.hpp/.cpp [skip ci]
